add logout

+++ src/redux/contacts/operations.js
@@ -1,10 +1,19 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { API } from "../auth/operations";
 
+const setAuthHeader = (token) => {
+  API.defaults.headers.common.Authorization = `Bearer ${token}`;
+};
+
 export const fetchContacts = createAsyncThunk(
   "contacts/fetchAll",
   async (_, thunkAPI) => {
     try {
+      const savedToken = thunkAPI.getState().auth.token;
+      if (!savedToken) {
+        return thunkAPI.rejectWithValue("Token doesn't exist");
+      }
+      setAuthHeader(savedToken);
       const response = await API.get("/contacts");
       return response.data;
     } catch (error) {
@@ -17,6 +26,11 @@ export const addContact = createAsyncThunk(
   "contacts/addContact",
   async (contact, thunkAPI) => {
     try {
+      const savedToken = thunkAPI.getState().auth.token;
+      if (!savedToken) {
+        return thunkAPI.rejectWithValue("Token doesn't exist");
+      }
+      setAuthHeader(savedToken);
       const response = await API.post("/contacts", contact);
       return response.data;
     } catch (error) {
@@ -29,6 +43,11 @@ export const deleteContact = createAsyncThunk(
   "contacts/deleteContact",
   async (contactId, thunkAPI) => {
     try {
+      const savedToken = thunkAPI.getState().auth.token;
+      if (!savedToken) {
+        return thunkAPI.rejectWithValue("Token doesn't exist");
+      }
+      setAuthHeader(savedToken);
       await API.delete(`/contacts/${contactId}`);
       return contactId;
     } catch (error) {
